requirements.txt
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
jinja2==3.1.2